+++ src/assets/javascripts/index.ts
@@ -86,6 +86,7 @@ import {
   SearchIndex
 } from "integrations"
 import {
+  patchCodeBlocks,
   patchTables,
   patchDetails,
   patchScrollfix,
@@ -178,6 +179,7 @@ export function initialize(config: unknown) {
 
   const keyboard$ = setupKeyboard()
 
+  patchCodeBlocks({ document$, viewport$ })
   patchDetails({ document$, hash$ })
   patchScripts({ document$ })
   patchSource({ document$ })
@@ -383,8 +385,11 @@ export function initialize(config: unknown) {
 
             /* Determine common prefix */
             let index = 0
-            while (a.charAt(index) === b.charAt(index))
-              index++
+            if (a === b)
+              index = a.length
+            else
+              while (a.charAt(index) === b.charAt(index))
+                index++
 
             /* Replace common prefix (i.e. base) with effective base */
             for (let i = 0; i < urls.length; i++)